requirements.txt
Django>=4.0,<5.0
djangorestframework
psycopg2-binary
django-rest-swagger
drf-yasg